+++ gatsby-node.js
@@ -1,9 +1,9 @@
+const {createPages} = require("./utils/create-pages")
+const {createSchema} = require("./utils/create-schema")
 const {onCreateNode} = require("./utils/on-create-node")
 const {sourceNodes} = require("./utils/source-nodes")
-const {
-  createSchemaCustomization,
-} = require("./utils/create-schema-customization")
 
-exports.createSchemaCustomization = createSchemaCustomization
+exports.createPages = createPages
+exports.createSchemaCustomization = createSchema
 exports.onCreateNode = onCreateNode
 exports.sourceNodes = sourceNodes